+++ scripts/serving_prod.sh
@@ -1,3 +1,3 @@
 #/bin/bash
 
-gunicorn -w 5 --threads 2 --bind 0.0.0.0:8000 --chdir /src/serving/ wsgi:app
+poetry run gunicorn -w 5 --threads 2 --bind 0.0.0.0:8000 plasticorigins.serving.wsgi:app
