+++ src/plasticorigins/serving/templates/results.html
@@ -1,0 +1,8 @@
+<table>
+    {% for key, value in result.items() %}
+    <tr>
+        <th> {{ key }} </th>
+        <td> {{ value }} </td>
+    </tr>
+    {% endfor %}
+</table>
