+++ src/plasticorigins/serving/templates/upload.html
@@ -1,0 +1,6 @@
+<!doctype html>
+<title>Upload new File</title>
+<h1>Upload new File</h1>
+<form method=post enctype=multipart/form-data> <input type=file name=file>
+    <input type=submit value=Upload>
+</form>
