+++ src/plasticorigins/serving/wsgi.py
@@ -1,0 +1,4 @@
+from plasticorigins.serving.app import app
+
+if __name__ == "__main__":
+    app.run()
