--- src/serving/templates/results.html
@@ -1,8 +1,0 @@
-<table>
-    {% for key, value in result.items() %}
-    <tr>
-        <th> {{ key }} </th>
-        <td> {{ value }} </td>
-    </tr>
-    {% endfor %}
-</table>
